requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.1
altgraph==0.17.5
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
boto3==1.42.32
botocore==1.42.32
celery==5.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
et_xmlfile==2.0.0
fast2sms==1.0.0
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
kombu==5.6.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.0
narwhals==2.15.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==3.0.0
pefile==2024.8.26
pillow==12.1.0
plotly==6.5.2
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.10.1
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
redis==7.1.0
reportlab==4.4.9
requests==2.32.5
s3transfer==0.16.0
schedule==1.2.2
sentry-sdk==2.50.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
tenacity==9.1.2
tkinterweb==4.16.5
tkinterweb-tkhtml==2.1.0
twilio==9.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.3.0
Werkzeug==3.1.5
yarl==1.22.0
keyring==24.1.0
PySide6
PySide6-Addons
httpx==0.27.0
APScheduler==3.10.1
ruff==0.28.0
mypy==1.9.0
psycopg[binary]==3.5.0
briefcase
